backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
qdrant-client==1.12.1
sentence-transformers==3.1.1
langgraph==0.2.53
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.19
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
numpy==1.26.4
